Moved TypeModal and VocabularyEditor JSX components to 'editor' subpackage
